Move PluginAssets class in proper place, use right class loader.
