+++ graylog2-server/src/main/java/org/graylog2/web/PluginAssets.java
@@ -14,11 +14,10 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.web.resources;
+package org.graylog2.web;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.graylog2.plugin.Plugin;
-import org.graylog2.web.ModuleManifest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
