+++ graylog2-server/src/main/java/org/graylog2/web/resources/AppConfigResource.java
@@ -17,7 +17,7 @@
 package org.graylog2.web.resources;
 
 import com.floreysoft.jmte.Engine;
-import org.apache.commons.io.IOUtils;
+import com.google.common.io.Resources;
 import org.graylog2.Configuration;
 
 import javax.inject.Inject;
@@ -26,7 +26,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import java.io.IOException;
-import java.net.URI;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -37,7 +38,8 @@ public class AppConfigResource {
 
     @Inject
     public AppConfigResource(Configuration configuration) throws IOException {
-        final String template = IOUtils.toString(ClassLoader.getSystemResourceAsStream("web-interface/config.js.template"));
+        final URL templateUrl = this.getClass().getResource("/web-interface/config.js.template");
+        final String template = Resources.toString(templateUrl, StandardCharsets.UTF_8);
         final Map<String, Object> model = new HashMap<String, Object>() {{
             put("serverUri", configuration.getRestTransportUri());
             put("appPathPrefix", "");
